Break the import cycles: move four things to the layer they belong to
Phase 3.1. Behaviour-preserving moves; the only logic change is that a
duplicated per-file mutex is retired in favour of the existing one.

decodeXmlEntitiesFromXppSource lived in the modify TOOL, and
utils/smartXmlBuilder.ts imported it from there — a three-module cycle
running the wrong way across the layer boundary, and the worst violation
the audit found. It is escapeXml's inverse, so it now sits beside it in
utils/xmlEscape.ts.

findD365FileOnDisk also lived in the modify tool, and four READ-ONLY
tools (tableInfo, enumInfo, queryInfo, viewInfo) imported it from there —
so reading a table pulled in the whole write path, its bridge adapter and
its op-spec registry. It answers "where does this metadata live", which
is a util's question. Now utils/objectFileLookup.ts.

ProjectFileManager and ProjectFileFinder lived in the create tool and
were imported by six other modules, including modify — which is what
closed the direct createD365File <-> modifyD365File cycle between the two
largest files in the codebase. Neither class is about creating an object;
both are about the .rnrproj, which is workspace state. Now
workspace/projectFile.ts, and their private withProjectFileLock is
dropped in favour of withFileLock from the atomic-write work rather than
carried along as a second per-file mutex.

createdArtifactLedger moved to workspace/ as well: it is session state
shared by create and undo, and leaving it under tools/ would have made
the new workspace/projectFile.ts import upward instead.

get_workspace_info was ~320 lines inline in the dispatcher's switch,
which made toolHandler.ts both the router and the largest single tool
implementation. Extracted to its own file like every other tool;
toolHandler drops 756 -> 426 lines.

tests/utils/layering.test.ts pins the direction so none of this can come
back. All four of its cases fail against the pre-fix tree.
