+++ src/tools/generateSmartForm.ts
@@ -13,7 +13,7 @@ import fs from 'fs';
 import { getConfigManager } from '../utils/configManager.js';
 import { defaultPackagesRoot } from '../utils/packagesRoot.js';
 import { resolveObjectPrefix, applyObjectPrefix, getObjectSuffix, applyObjectSuffix } from '../utils/modelClassifier.js';
-import { ProjectFileManager } from './createD365File.js';
+import { ProjectFileManager } from '../workspace/projectFile.js';
 import { extractModelFromProject, findProjectInSolution } from '../utils/projectUtils.js';
 import { normalizeD365Xml } from '../utils/d365XmlNormalizer.js';
 import { validateFormPatternXml } from '../validation/formPatternValidator.js';
