+++ src/utils/xmlEscape.ts
@@ -39,3 +39,41 @@ export function escapeXml(value: unknown): string {
 export function escapeXmlAttr(value: unknown): string {
   return escapeXml(value).replace(/"/g, '&quot;');
 }
+
+/**
+ * Decode the standard XML entities (&lt;, &gt;, &apos;, &quot;, &amp;) and
+ * normalise line endings by stripping xml2js's &#xD; representation of carriage
+ * return.
+ *
+ * `&amp;` is decoded LAST so that a sequence like `&amp;quot;` first becomes
+ * `&quot;` and can then be decoded to `"` — decoding it first would
+ * double-unescape.
+ */
+function decodeStandardXmlEntities(source: string): string {
+  return source
+    // xml2js Builder escapes \r as &#xD; — strip it to normalise to LF-only line endings
+    .replace(/&#xD;/g, '')
+    .replace(/&lt;/g, '<')
+    .replace(/&gt;/g, '>')
+    .replace(/&apos;/g, "'")
+    .replace(/&quot;/g, '"')
+    .replace(/&amp;/g, '&');
+}
+
+/**
+ * Decode XML entities from X++ source code.
+ *
+ * X++ source should never contain entity-encoded characters — `/// <summary>`
+ * doc comments, generic types like `List<str>`, and comparison operators like
+ * `x < y` all use literal `<` and `>`. When an AI model copies code from an
+ * SSRS report's entity-encoded <Text> block and passes it as `methodCode`, the
+ * entities would otherwise survive into the CDATA section and corrupt the source.
+ *
+ * Lives here, beside escapeXml, because it is that function's inverse — and
+ * because its previous home was the modify TOOL, which made src/utils/
+ * smartXmlBuilder.ts import from src/tools/ to reach it: a three-module cycle
+ * running the wrong way across the layer boundary.
+ */
+export function decodeXmlEntitiesFromXppSource(source: string): string {
+  return decodeStandardXmlEntities(source);
+}
